# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem RW API Documentation",
        "description": "Sistem RW OpenApi description",
        "contact": {
            "email": "achmad.zuhri@varx.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": ["Auth"],
                "summary": "Login function",
                "description": "Login to app",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/dashboard": {
            "get": {
                "tags": ["Dashboard"],
                "summary": "Get Dashboard function",
                "description": "Get all data dashboard ",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/logout": {
            "get": {
                "tags": ["Auth"],
                "summary": "Logout from App",
                "description": "Logout from app",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/jimpitan/tampil": {
            "post": {
                "tags": ["Jimpitan"],
                "summary": "Get Jimpitan data function",
                "description": "Get all jimpitan data",
                "operationId": "s",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dari": {
                                        "type": "string"
                                    },
                                    "ke": {
                                        "type": "string"
                                    },
                                    "cari": {
                                        "type": "string"
                                    },
                                    "kategori": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "dari": "2022-12-01",
                                    "ke": "2022-12-30",
                                    "cari": "",
                                    "kategori": "harian"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/jimpitan/tambah": {
            "post": {
                "tags": ["Jimpitan"],
                "summary": "Add Jimpitan function",
                "description": "Adding Jimpitan data to sistem",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "no_kk": {
                                        "type": "number"
                                    },
                                    "nominal": {
                                        "type": "number"
                                    },
                                    "kategori": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "no_kk": 3471072108961377,
                                    "nominal": 500,
                                    "kategori": "harian"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/jimpitan/ubah/{id}": {
            "post": {
                "tags": ["Jimpitan"],
                "summary": "Update Jimpitan data function",
                "description": "Updating jimpitan data by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nominal": {
                                        "type": "number"
                                    },
                                    "kategori": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nominal": 500,
                                    "kategori": "harian"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/jimpitan/hapus/{id}": {
            "get": {
                "tags": ["Jimpitan"],
                "summary": "Delete Jimpitan data",
                "description": "Deleting Jimpitan data by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "A short description for oauth2 security ",
                "scheme": "bearer",
                "in": "header",
                "bearerAuth": "bearerAuth",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "passport": []
        }
    ]
}